% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lights.R
\name{get_state}
\alias{get_state}
\title{Get's the current state of your Elgato Key Light}
\usage{
get_state(keylight_ip = NULL)
}
\arguments{
\item{keylight_ip}{optional IP address of your Keylight. Setting the EKL_IP
env var is preferred}
}
\description{
Get's the current state of your Elgato Key Light
}
