% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lights.R
\name{set_state}
\alias{set_state}
\title{Fine grained control of the light state}
\usage{
set_state(state, brightness = NA, temp = NA, keylight_ip = NULL)
}
\arguments{
\item{state}{One of "on" or "off"}

\item{brightness}{int from 1-100. If not supplied the current setting is retained.}

\item{temp}{int corresponding to the desired colour temperature (143-344)
If not supplied the current setting is retained.}

\item{keylight_ip}{optional IP address of your Keylight. Setting the EKL_IP
env var is preferred}
}
\description{
Fine grained control of the light state
}
