% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lights.R
\name{toggle}
\alias{toggle}
\title{Toggles the light on/off}
\usage{
toggle(keylight_ip = NULL)
}
\arguments{
\item{keylight_ip}{optional IP address of your Keylight. Setting the EKL_IP
env var is preferred}
}
\description{
Toggles the light on/off
}
